refactor(dsh-mcp-manager): 阶段2收敛——执行管道域 src/pipeline/ 成形（interface.ts 首个应用）
Summary:
- 新建 src/pipeline/{args,msg,redact,timeout,authorize,project}.ts + interface.ts：
  执行管道纯函数族自 middleware-utils.ts / call-result.ts 迁入（normalizeArguments
  / msgOf / createRedactor / withTimeout / globMatch / projectCallToolResult /
  defaultCallResultFallbackText + 类型面），零行为变更（B14 已修语义原样保留）
- pipeline/interface.ts = 目录唯一对外引用面（D10 首个应用），middleware.ts /
  supervisor.ts / middleware-register.ts / index.ts / middleware-utils.ts 跨目录
  引用全部改走 interface.ts；verify-dir-imports 首个真实生效（PASS）
- call-result.ts 删除（内容全量迁 pipeline/project.ts）；middleware-utils.ts
  收敛为 workspace/catalog 域函数（阶段 4/5 迁出）；middleware.ts re-export 块
  拆分（pipeline 函数经 interface.ts 转发，lib 导出面不变）
- protocol.ts 注释同步模块路径；middleware-utils 头部注释按现状修订

Design decisions:
- policy 裁决族（policyAllows/isToolDenied 等）依赖 workspace 域函数
  （parseFullServerName/MIDDLEWARE_GLOBAL_ROOT，阶段 4 迁入），本阶段不强行迁入，
  避免跨目录引用破坏单向依赖；authorize.ts 仅含自包含 globMatch
- 不建 executeTool 大组合器：pipeline 保持单文件单职责纯函数族，两路径同构
  由契约测试强制（下一 commit）

Related issue: #664
